Add files via upload
This is my submission for the Module 2 Challenge Assignment comparing Multiple Year Stock Data, summarizing the data, and then extracting the Ticker symbol with the greatest percentage increase, the greatest percentage decrease, and greatest total volume.

Each VB file corresponds to one of two buttons located on each of the worksheets in my workbook which will run the subroutines.  These subroutines then apply to all sheets at once.  This is for convenience allowing the user to clear the data and re-run the calculations from any of the four worksheets.